requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
SQLAlchemy==2.0.27
langchain==0.1.8
langchain-openai==0.0.6
python-dotenv==1.0.1
faiss-cpu==1.7.4
tiktoken==0.6.0
httpx==0.26.0
python-multipart
aiofiles
chromadb